app/pyproject.toml
[project]
name = "Spiro"
version = "0.1.1"
dependencies = [
    "flet_audio_recorder",
    "numpy"
]